node: New package extracted from "ovn"
We had few code blocks scattered across "ovn" package (and others)
that dealt with identifying running services on individual nodes in
the cluster.

Since these functions do not require any internal fucntionality from
"ovn" package, and they are unversally useful, They are extracted to
a new pacakge called "node".
